Moving "AD" logic to symoblic function
I had mistakenly developed symbolic differentiation.  {sigh}  This
pulls that out, and puts it in a separately function.

The output model diagnostics are the same, and on the up-side I
added a diff() method to expressions (mimicing the sympy diff API).

But for now, the AD logic has been removed/disabled.